Move useIsDraggingWithin to a shared hook
