+++ packages/block-library/src/navigation-link/shared/index.js
@@ -16,3 +16,4 @@ export { useHandleLinkChange } from './use-handle-link-change';
 export { useIsInvalidLink } from './use-is-invalid-link';
 export { InvalidDraftDisplay } from './invalid-draft-display';
 export { useEnableLinkStatusValidation } from './use-enable-link-status-validation';
+export { useIsDraggingWithin } from './use-is-dragging-within';
